Cache Github API calls to disk
Addresses [https://github.com/telia-oss/github-pr-resource/issues/142]

When this resource is configured to filter versions, it can sometimes
result in a large number of repeated calls to the github API to
determine whether a PR should be emitted as a version or not.
In extreme cases, this can result in usage that exceeds Github's API
rate limit, causing the resource to be unavailable until the rate limit
resets.

This change mitigates the issue by caching the results of Github API
calls to disk, such that repeated calls to fetch the modified file list
of Pull Requests which haven't updated since the previous call, will
instead be served from the cache.

As Concourse intentionally re-uses containers for Check processes over
a long time period, this cache should be long-lived in most deployments.
[https://concourse-ci.org/implementing-resource-types.html#resource-check]

It should be noted that writing a cached response to disk can fail.
The underlying library, peterbourgon/diskv, returns a result in these
cases. However, the http caching library, gregjones/httpcache, silently
swallows any such errors. If a permission issue continually prevents
writing cached responses to disk, the Check will still proceed as normal,
and this resource will behave as if caching wasn't implemented.