Add positive and negative head branch filters
For some workflows, it can be useful to treat pull requests differently
depending on the name of the head branch for the pull request.

For example, a team following scaled trunk-based development might use
the `release/*` naming convention for release branches, and may have
different build/deploy processes for those release branches.

This adds `head_branches` and `ignore_head_branches` settings to this
resource, specified as a glob pattern to align with similar filters on
the Concourse git resource, so that teams can run different jobs based
on the head branch of a PR.

See: https://trunkbaseddevelopment.com/#scaled-trunk-based-development