Filter ignore_paths before wanted paths
This switches the order in which modified files are checked on pull
requests. Currently ignored and wanted paths are checked independently
of each other which can lead to new resource versions being created
despite paths being ignored.

If I have the following config: `paths: app-a/, ignore_paths:
app-a/config/` and modified files `app-a/config/settings.ini,
app-b/files.ext`. The check logic determines `app-a/config/settings.ini`
is wanted and doesn't skip the Loop, which is OK and expected.

However the `FilterIgnorePath` check removes the `settings.ini` file as
un-matched but sees the `app-b/files.ext` and concludes that it must be
wanted. A new resource is created from a file that doesn't really match
the wanted source paths.

This commit switches the order of those checks and removes the files
that match the `ignore_paths` from the list of files that are then
checked against the `paths`. It also updates the test to check this
condition.

Lastly it adds a new function `createGoodTestPRWithCount()` to return
a slice of pull requests that will not cause the Check loop to silently
bail on a label, skipCI flag or other. I found that some test cases were
being skipped because the modified file lists are assigned to pull
requests that get skipped.