Add full wildcard globbing for nested paths
This adds full wildcard globbing for paths with standard operators like
`**` as [documented in the gobwas/glob library][docs]. This builds on
the standard Go library implentation of globbing which does not make use
of the more complex operations.

For example if you wanted to trigger pull request checks on any files
with the extensions `.tf` you would have to create a paths config for
different levels of nesting.

```yaml
source:
    paths:
    - */*/*.tf
    - */*/*/*.tf
```

This is inflexible and probably not expected behaviour. With this commit
you can set paths such as `**.tf` that will match all files on a pull
request with the `.tf` extension. The same behaviour is implemented for
`ignore_paths`.

[docs]: https://github.com/gobwas/glob/tree/e7a84e9525fe90abcda167b604e483cc959ad4aa#example