Handle situation when `ignored_paths` and `paths` are set
Currently, the behaviour of `ignored_paths` applies to ALL the changed
files in the PR. You cannot apply the `ignored_files` filter to only
those files in the selected `paths`. This limits the usefulness of the
`ignored_paths`.

What we want is : 'Trigger the build when there are files in `paths` that
are not also in `ignored_paths`'